rename lib
