+++ exporters/committime/values.yaml
@@ -1,0 +1,7 @@
+# Values file for exporter helm chart
+app_name: committime-exporter
+env_from_secrets: 
+- github-secret
+source_context_dir: exporters/committime
+source_ref: refactor-leadtime
+source_url: https://github.com/etsauer/mdt-quickstart.git
